Using AWS Credentials Chain
This solves https://github.com/SumoLogic/sumologic-kinesis-connector/issues/12

It now uses the AWSCredentialsProviderChain instead of only looking in a file.  This allows using ENV vars, or roles on EC2 and ECS instances.
This is the same flow as the Default Chain except that as the final link, it looks in the SumologicConnector.properties file.

I also removed some unused import statements